添加

+++ HaiYin/app/src/main/java/com/haiyin/gczb/my/entity/GetMyContractProjectsEntity.java
@@ -35,7 +35,7 @@ public class GetMyContractProjectsEntity extends BaseEntity {
         private String projectId;
         private String title;
         private String cityName;
-        private int amount;
+        private Double amount;
         private String summary;
         private int days;
         private String address;
@@ -64,11 +64,11 @@ public class GetMyContractProjectsEntity extends BaseEntity {
             this.cityName = cityName;
         }
 
-        public int getAmount() {
+        public Double getAmount() {
             return amount;
         }
 
-        public void setAmount(int amount) {
+        public void setAmount(Double amount) {
             this.amount = amount;
         }
