+++ HaiYin/app/src/main/java/com/haiyin/gczb/order/entity/OrderListsEntity.java
@@ -38,7 +38,7 @@ public class OrderListsEntity extends BaseEntity {
         private String projectId;
         private String title;
         private String cityName;
-        private int amount;
+        private Double amount;
         private String summary;
         private int days;
         private String address;
@@ -69,11 +69,11 @@ public class OrderListsEntity extends BaseEntity {
             this.cityName = cityName;
         }
 
-        public int getAmount() {
+        public Double getAmount() {
             return amount;
         }
 
-        public void setAmount(int amount) {
+        public void setAmount(Double amount) {
             this.amount = amount;
         }
